COAR NotifyService IT failures: move failings to a separate class to see what happens
