+++ dspace-server-webapp/src/test/java/org/dspace/app/rest/NotifyServiceRestRepositoryIT.java
@@ -1300,70 +1300,6 @@ public class NotifyServiceRestRepositoryIT extends AbstractControllerIntegration
     }
 
     @Test
-    public void NotifyServiceInboundPatternConstraintRemoveOperationTest() throws Exception {
-
-        context.turnOffAuthorisationSystem();
-
-        NotifyServiceEntity notifyServiceEntity =
-            NotifyServiceBuilder.createNotifyServiceBuilder(context, "service name")
-                                .withDescription("service description")
-                                .withUrl("https://service.ldn.org/about")
-                                .withLdnUrl("https://service.ldn.org/inbox")
-                                .build();
-
-        context.restoreAuthSystemState();
-
-        List<Operation> ops = new ArrayList<Operation>();
-        AddOperation inboundAddOperationOne = new AddOperation("notifyServiceInboundPatterns/-",
-            "{\"pattern\":\"patternA\",\"constraint\":\"itemFilterA\",\"automatic\":\"false\"}");
-
-        AddOperation inboundAddOperationTwo = new AddOperation("notifyServiceInboundPatterns/-",
-            "{\"pattern\":\"patternB\",\"constraint\":\"itemFilterB\",\"automatic\":\"true\"}");
-
-        ops.add(inboundAddOperationOne);
-        ops.add(inboundAddOperationTwo);
-        String patchBody = getPatchContent(ops);
-
-        String authToken = getAuthToken(admin.getEmail(), password);
-        getClient(authToken)
-            .perform(patch("/api/ldn/ldnservices/" + notifyServiceEntity.getID())
-                .content(patchBody)
-                .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.notifyServiceInboundPatterns", hasSize(2)))
-            .andExpect(jsonPath("$",
-                allOf(
-                    matchNotifyService(notifyServiceEntity.getID(), "service name", "service description",
-                        "https://service.ldn.org/about", "https://service.ldn.org/inbox"),
-                    hasJsonPath("$.notifyServiceInboundPatterns", containsInAnyOrder(
-                        matchNotifyServicePattern("patternA", "itemFilterA", false),
-                        matchNotifyServicePattern("patternB", "itemFilterB", true)
-                    ))
-                )));
-
-        RemoveOperation inboundRemoveOperation = new RemoveOperation("notifyServiceInboundPatterns[1]/constraint");
-        ops = new ArrayList<Operation>();
-        ops.add(inboundRemoveOperation);
-        patchBody = getPatchContent(ops);
-
-        getClient(authToken)
-            .perform(patch("/api/ldn/ldnservices/" + notifyServiceEntity.getID())
-                .content(patchBody)
-                .contentType(MediaType.APPLICATION_JSON_PATCH_JSON))
-            .andExpect(status().isOk())
-            .andExpect(jsonPath("$.notifyServiceInboundPatterns", hasSize(2)))
-            .andExpect(jsonPath("$",
-                allOf(
-                    matchNotifyService(notifyServiceEntity.getID(), "service name", "service description",
-                        "https://service.ldn.org/about", "https://service.ldn.org/inbox"),
-                    hasJsonPath("$.notifyServiceInboundPatterns", containsInAnyOrder(
-                        matchNotifyServicePattern("patternA", "itemFilterA", false),
-                        matchNotifyServicePattern("patternB", null, true)
-                    ))
-                )));
-    }
-
-    @Test
     public void NotifyServiceInboundPatternConstraintRemoveOperationBadRequestTest() throws Exception {
 
         context.turnOffAuthorisationSystem();
